+++ transforms/cpp-distance.h
@@ -1,2 +1,2 @@
-double *CppTwoDimensionalDistanceTransform(long *segmentation, long resolution[3]);
-long *CppDilateData(long *data, long resolution[3], double distance);
+float *CppTwoDimensionalDistanceTransform(long *segmentation, long resolution[3]);
+long *CppDilateData(long *data, long resolution[3], float distance);
